+++ assets/migrations/azure/005_add_conditions_to_tuples.sql
@@ -1,6 +1,6 @@
 -- +goose Up
-ALTER TABLE tuple ADD condition_name VARCHAR(256), condition_context VARBINARY(MAX);
-ALTER TABLE changelog ADD condition_name VARCHAR(256), condition_context VARBINARY(MAX);
+ALTER TABLE tuple ADD condition_name VARCHAR(256) COLLATE Latin1_General_BIN2, condition_context VARBINARY(MAX);
+ALTER TABLE changelog ADD condition_name VARCHAR(256) COLLATE Latin1_General_BIN2, condition_context VARBINARY(MAX);
 
 -- +goose Down
 ALTER TABLE tuple DROP COLUMN condition_name, condition_context;