+++ cmd/validatemodels/validate_models.go
@@ -10,6 +10,7 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/openfga/openfga/pkg/storage"
+	"github.com/openfga/openfga/pkg/storage/azure"
 	"github.com/openfga/openfga/pkg/storage/mysql"
 	"github.com/openfga/openfga/pkg/storage/postgres"
 	"github.com/openfga/openfga/pkg/storage/sqlcommon"
@@ -67,6 +68,8 @@ func runValidate(_ *cobra.Command, _ []string) error {
 		db, err = postgres.New(uri, cfg)
 	case "sqlite":
 		db, err = sqlite.New(uri, cfg)
+	case "azure":
+		db, err = azure.New(uri, cfg)
 	case "":
 		return fmt.Errorf("missing datastore engine type")
 	case "memory":